// webview-ui/src/thor/model/Product.ts
export type Product = {
	/**
	 *
	 * @type {string}
	 * @memberof Product
	 */
	name: string
	/**
	 * the base price of the product before features
	 * @type {number}
	 * @memberof Product
	 */
	price: number
	/**
	 * Availability status of the product.
	 * @type {string}
	 * @memberof Product
	 */
	status: ProductStatusEnum
	/**
	 *
	 * @type {string}
	 * @memberof Product
	 */
	description?: string
	/**
	 * the sale price of the product if any
	 * @type {number}
	 * @memberof Product
	 */
	salePrice?: number
	/**
	 * the duration of the product if applicable
	 * @type {number}
	 * @memberof Product
	 */
	duration?: number
	/**
	 *
	 * @type {string}
	 * @memberof Product
	 */
	type?: ProductTypeEnum
	/**
	 *
	 * @type {Array<ProductFeature>}
	 * @memberof Product
	 */
	features?: Array<ProductFeature>
	/**
	 * Unique identifier for object in the system
	 * @type {string}
	 * @memberof Product
	 */
	id?: string
	/**
	 * UUID of owner of the object in the system
	 * @type {string}
	 * @memberof Product
	 */
	ownerId?: string
	/**
	 * Date of object creation
	 * @type {Date}
	 * @memberof Product
	 */
	createdDate?: Date
	/**
	 * Data, including hash of the key(s) used to encrypt this record.
	 * @type {string}
	 * @memberof Product
	 */
	keyHash?: string
	/**
	 * Last user to access object
	 * @type {string}
	 * @memberof Product
	 */
	lastAccessedById?: string
	/**
	 * Timestamp of last access of object
	 * @type {Date}
	 * @memberof Product
	 */
	lastAccessedDate?: Date
	/**
	 * Unique identifier for user who last modifed the object in the system
	 * @type {string}
	 * @memberof Product
	 */
	lastModifiedById?: string
	/**
	 * Date of last object modification
	 * @type {Date}
	 * @memberof Product
	 */
	lastModifiedDate?: Date
}
// ...
export enum ProductStatusEnum {
	PENDING = "pending",
	SALE = "sale",
	CLEARANCE = "clearance",
	AVAILABLE = "available",
	DISCONTINUED = "discontinued",
	OUTOFSTOCK = "out_of_stock",
}
/**
 * @export
 * @enum {string}
 */
export enum ProductTypeEnum {
	PHYSICAL = "physical",
	CLOUD = "cloud",
	DOWNLOAD = "download",
	SERVICE = "service",
	MEDIA = "media",
	PERFORMANCE = "performance",
	OTHER = "other",
}